# trader_interactive/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/listings": {
            "get": {
                "tags": [
                    "listings"
                ],
                "summary": "Get list of listings",
                "description": "Returns list of listings",
                "operationId": "getListings",
                "responses": {
                    "200": {
                        "description": "Listing Details Found"
                    },
                    "500": {
                        "description": "Listing Details Not Found"
                    }
                }
            }
        },
        "/api/listing-details/id={id}": {
            "get": {
                "tags": [
                    "listing-details"
                ],
                "summary": "Get Individual Listing Details",
                "description": "Returns Individual Listing Details",
                "operationId": "getListingDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "listing id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Individual Listing Details Found"
                    },
                    "500": {
                        "description": "Individual Listing Details Not Found"
                    }
                }
            }
        },
        "/api/search-listing-details/make={make}/model={model}/min_price={min_price}/max_price={max_price}": {
            "get": {
                "tags": [
                    "search-listing-details"
                ],
                "summary": "Get Searched Listing Details",
                "description": "Returns Searched Listing Details",
                "operationId": "getUserListings",
                "parameters": [
                    {
                        "name": "make",
                        "in": "path",
                        "description": "make",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "model",
                        "in": "path",
                        "description": "model name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "min_price",
                        "in": "path",
                        "description": "minimum price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "max_price",
                        "in": "path",
                        "description": "maximum price",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Searched Listing Details Found"
                    },
                    "500": {
                        "description": "Searched Listing Details Not Found"
                    }
                }
            }
        },
        "/api/create-listings": {
            "post": {
                "tags": [
                    "create-listings"
                ],
                "summary": "Create listings",
                "description": "Create listings",
                "operationId": "createListings",
                "parameters": [
                    {
                        "name": "make",
                        "in": "formData",
                        "description": "make",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "model",
                        "in": "formData",
                        "description": "model",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "year",
                        "in": "formData",
                        "description": "year",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "colour",
                        "in": "formData",
                        "description": "colour",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "thumbnail_1",
                        "in": "formData",
                        "description": "thumbnail_1",
                        "required": true,
                        "schema": {
                            "type": "file"
                        },
                        "type": "file"
                    },
                    {
                        "name": "thumbnail_2",
                        "in": "formData",
                        "description": "thumbnail_2",
                        "required": true,
                        "schema": {
                            "type": "file"
                        },
                        "type": "file"
                    },
                    {
                        "name": "thumbnail_3",
                        "in": "formData",
                        "description": "thumbnail_3",
                        "required": true,
                        "schema": {
                            "type": "file"
                        },
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listings Created Successfully"
                    },
                    "500": {
                        "description": "Listings Not Created"
                    }
                }
            }
        },
...
    },
...
}